+++ main.go
@@ -18,6 +18,7 @@ import (
 	"github.com/bitrise-io/go-utils/log"
 	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/bitrise-io/go-utils/sliceutil"
+	"github.com/bitrise-steplib/steps-virtual-device-testing-for-android/step"
 	toolresults "google.golang.org/api/toolresults/v1beta3"
 )
 
@@ -136,7 +137,7 @@ func main() {
 					failf("Failed to write in tabwriter, error: %s", err)
 				}
 
-				successful, err = GetSuccessOfExecution(responseModel.Steps)
+				successful, err = step.GetSuccessOfExecution(responseModel.Steps)
 				if err != nil {
 					failf("Failed to process results, error: %s", err)
 				}
@@ -358,37 +359,4 @@ func processStepResult(step *toolresults.Step) string {
 		outcome = colorstring.Blue(outcome)
 	}
 	return outcome
-}
-
-func GetSuccessOfExecution(steps []*toolresults.Step) (bool, error) {
-	outcomeByDimension, err := getOutcomeByDimension(steps)
-	if err != nil {
-		return false, err
-	}
-
-	for _, outcome := range outcomeByDimension {
-		if outcome.Summary != "success" {
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-func getOutcomeByDimension(steps []*toolresults.Step) (map[string]*toolresults.Outcome, error) {
-	groupedByDimension := make(map[string]*toolresults.Outcome)
-	for _, step := range steps {
-		key, err := json.Marshal(step.DimensionValue)
-		if err != nil {
-			return nil, err
-		}
-		if key != nil {
-			dimensionStr := string(key)
-			if groupedByDimension[dimensionStr] == nil || groupedByDimension[dimensionStr].Summary != "success" {
-				groupedByDimension[dimensionStr] = step.Outcome
-			}
-		}
-	}
-
-	return groupedByDimension, nil
 }
